Create module, move relevant methods
